refactor: OAuth app directory strcuture
